Rename tables to something more generic
